requirements.txt
python-pptx
bs4
tenacity
openai
tiktoken